STY: Reviewer comments, move function.
